gnn/requirements.txt
humemai==1.1.2
room-env==2.0.7
numpy==1.26.4
torch==2.1.1
torch-scatter==2.1.2
torch_geometric==2.4.0
tqdm==4.66.4
PyYAML>=6.0.1
ipython
ipykernel